requirements.txt
speedtest-cli
requests
python-dotenv
